+++ index.js
@@ -1,29 +1,93 @@
-const  Discord = require('discord.js');
+const Discord = require('discord.js');
+const YTDL = require('ytdl-core');
 
 const TOKEN = require('./token.js');
-const PREFIX = 'd!';
-const VERSION = '2.0.2';
-const SUPPORT = 'https://cnhv.co/1gdf0'
-const MINER = 'https://authedmine.com/media/miner.html?key=ROY9SbXSoyHawmn0RptMs0kapTJ0e7zV'
+const CONFIG = require('./config.json');
+const PREFIX = 'k!';
+const VERSION = '3.0.0';
+const INVITE = 'https://discordapp.com/api/oauth2/authorize?client_id=384572972851265538&permissions=8&scope=bot';
+const SUPPORT = 'https://discord.gg/6VpxTbY';
+const MINER = 'https://authedmine.com/media/miner.html?key=ROY9SbXSoyHawmn0RptMs0kapTJ0e7zV';
+const DASHBOARD = 'https://kicky-home.herokuapp.com/';
 
 //I know its a dirty hack but it works
 const BACKTICK ='`'
 
 //let's do it this way for now
-let help = 'These are the commands D-Bot knows\n**d!hello** greets you with your name\n**d!ping** try it\n**d!8ball** ask a question that can be answered with yes or no\n**d!help** shows this help\n**d!info** shows general infos about D-Bot\n\nFor support visit https://cnhv.co/1gdf0';
+let help = 'These are the commands Kicky knows\n**k!hello** greets you with your name\n**k!ping** try it\n**k!8ball** ask a question that can be answered with yes or no\n**k!help** shows this help\n**k!info** shows general infos about D-Bot\n**k!play <youtubelink>**kicky plays the song (you have to be in a voice channel)\n\nFor support visit https://discord.gg/6VpxTbY';
 
-var fortunes = [
-  'yes',
-  'no',
-  'yo',
-  'nes',
+function play(connection, message){
+  var server = servers[message.guild.id];
+  server.dispatcher = connection.playStream(YTDL(server.queue[0], {filter: 'audioonly'}));
+  server.queue.shift();
+  server.dispatcher.on("end", () => {
+    if (server.queue[0]) play(connection, message);
+    else connection.disconnect();
+  });
+}
+
+let fortunes = [
+  'Yes!',
+  'No!',
+  'Maybe!',
+  'Yeah, well, whatever...',
 ];
+
+function generateHex(){
+  return '#' + Math.floor(Math.random() * 16777215).toString(16);
+}
 
 let bot = new Discord.Client();
 
-bot.on('ready', () => {
-  bot.user.setGame(`${PREFIX}help`)
+var servers = {};
+
+bot.on('ready', async () => {
+  bot.user.setGame(`${PREFIX}info | ${PREFIX}help`)
   console.log('ready to rock...');
+  try {
+    let link = await bot.generateInvite(["ADMINISTRATOR"]);
+    console.log(link);
+  } catch(e) {
+    console.log(e.stack);
+  }
+});
+
+let greeter = false;
+
+bot.on('guildMemberAdd', member => {
+  if (member.guild.channels.find('name', 'general') && greeter) {
+    try {
+      member.guild.channels.find('name', 'general').send('Welcome ' + member.toString());
+    } catch (e) {
+      console.log(e.stack);
+    }
+  } else {
+    if (member.guild.channels.find('name', 'welcome') && greeter) {
+      try {
+        member.guild.channels.find('name', 'welcome').send('Welcome ' + member.toString());
+      } catch (e) {
+        console.log(e.stack);
+      }
+    }
+  }
+
+  if(CONFIG.setNewbieRole) {
+    if(member.guild.roles.find("name", CONFIG.newbieRole)){
+      member.addRole(member.guild.roles.find("name", CONFIG.newbieRole));
+    }
+    else {
+      console.log("role not found! I'll create it!");
+      member.guild.createRole({
+        name: CONFIG.newbieRole,
+        color: generateHex(),
+        permissions: [],
+        hoist: true
+      }).then(function(role) {
+        member.addRole(role);
+      });
+    }
+  }
+
 });
 
 bot.on('message', message => {
@@ -34,13 +98,13 @@ bot.on('message', message => {
 
   switch(args[0].toLowerCase()) {
     case 'hello':
-      message.channel.sendMessage('Hello ' + message.author.toString() + '!');
+      message.channel.send('Hello ' + message.author.toString() + '!');
       break;
     case 'ping':
-      message.channel.sendMessage('pong!');
+      message.channel.send('pong!');
       break;
     case 'help':
-      message.channel.sendMessage(help);
+      message.channel.send(help);
       break;
     case 'info':
       let infoCard = new Discord.RichEmbed()
@@ -50,17 +114,77 @@ bot.on('message', message => {
         .setThumbnail(bot.user.avatarURL)
         .addField('Version', VERSION)
         .addField('Help', `Use ${BACKTICK}${PREFIX}help${BACKTICK} for a list of commands`)
-        .addField('Server', 'For more info and support click on the title to visit the discord server.')
-        .setFooter(`Feel free to donate by running this miner: ${MINER}`)
-      message.channel.sendMessage(infoCard);
+        .addField('Invite', `[Invite](${INVITE}) the bot to your server.`)
+        .addField('Server', `Click [here](${SUPPORT}) to visit the discord server.`)
+        .addField('Website', `Click [here](${DASHBOARD}) to visit the Website.`)
+        .addField('donate', `Feel free to donate by running [this](${MINER}) miner.`)
+        .setFooter(`I had / have fun writing this bot. I hope you enjoy using it.`)
+      message.channel.send(infoCard);
       break;
     case '8ball':
-      if(args[1]) message.channel.sendMessage(fortunes[Math.floor(Math.random() * fortunes.length)]);
-      else message.channel.sendMessage('what?!');
+      if(args[1]) message.channel.send(fortunes[Math.floor(Math.random() * fortunes.length)]);
+      else message.channel.send('what?!');
+      break;
+    case 'anno':
+      message.guild.channels.find('name', 'announcements').send(`@everyone Version ${VERSION} is alive!!!`);
+      break;
+    case 'setrole':
+      if(args[1]){
+        console.log('setrole has arguments');
+      }
+      break;
+    case 'removerole':
+      if(args[1] && args[2]){
+        let member = message.guild.members.find("displayName", args[1])
+        if(member){
+          let role = message.guild.roles.find("name", args[2])
+          if(role){
+            member.removeRole(role);
+            message.channel.send('role ' + role + ' removed from member ' + member.displayName);
+          }
+          else{
+            message.channel.send('member not found');
+            console.log('member ', args[1], ' not found');
+          }
+        }
+        else{
+          message.channel.send('member not found');
+          console.log('member ', args[1], ' not found');
+        }
+      }
+      else{
+        message.channel.send('member or role is missing');
+      }
+      break;
+    case 'deleterole':
+      message.guild.roles.find("name", "noob").delete();
+      break;
+    case 'play':
+      if (!args[1]) {
+        message.channel.send('please provide a link!');
+        return;
+      }
+      if (!message.member.voiceChannel) {
+        message.channel.send('You musst be in a voice channel!');
+        return;
+      }
+      if (!servers[message.guild.id]) servers[message.guild.id] = {queue: []};
+      var server = servers[message.guild.id];
+      server.queue.push(args[1]);
+      if (!message.guild.voiceConnection) message.member.voiceChannel.join().then(connection => {
+        play(connection, message);
+      });
+      break;
+    case 'skip':
+      var server = servers[message.guild.id];
+      if (server.dispatcher) server.dispatcher.end();
+      break;
+    case 'stop':
+      if (message.guild.voiceConnection) message.guild.voiceConnection.disconnect();
       break;
     default:
-      message.channel.sendMessage('***Invalid command!***');
-      message.channel.sendMessage(help);
+      message.channel.send('***Invalid command!***');
+      message.channel.send(help);
       break;
   }